web view functions moved to utils with async changes
