+++ src/webView/utils.ts
@@ -1,44 +1,38 @@
 import * as vscode from 'vscode';
 import * as fs from "fs";
 import { config } from "../config";
-import { NewComment } from '../extension';
+import { NewComment, commentCache } from '../extension';
 
 export function keepComment(commentText: string) {
 	// from comment text find line number
-	const content = fs.existsSync(config.commentJSONPath)
-          ? JSON.parse(fs.readFileSync(config.commentJSONPath, "utf-8"))
-          : {};
+	const content = commentCache.get(config.commentJSONPath) || {};
 	for (const key in content) {
 		if (content[key] === commentText) {
 			const lineNumber = +key.split('-')[0];
-			const codeLine = config.document?.lineAt(lineNumber-1).text || "";
+			const codeLine = config.document?.lineAt(lineNumber - 1).text || "";
 			const newKey = lineNumber + "-" + btoa(codeLine);
 			const value = content[key];
 			delete content[key];
 			content[newKey] = value;
-			fs.writeFileSync(
-				config.commentJSONPath,
-				JSON.stringify(content, null, 2)
-			);
+			fs.promises.writeFile(config.commentJSONPath,
+				JSON.stringify(content, null, 2));
 			break;
 		}
 	}
 }
 
 export function deleteComment(commentText: string) {
-	const content = fs.existsSync(config.commentJSONPath)
-          ? JSON.parse(fs.readFileSync(config.commentJSONPath, "utf-8"))
-          : {};
-    for (const key in content) {
-      if (content[key] === commentText) {
-        delete content[key];
-        fs.writeFileSync(
-          config.commentJSONPath,
-          JSON.stringify(content, null, 2)
-        );
-        break;
-      }
-    }
+	const content = commentCache.get(config.commentJSONPath) || {};
+	for (const key in content) {
+		if (content[key] === commentText) {
+			delete content[key];
+			fs.promises.writeFile(
+				config.commentJSONPath,
+				JSON.stringify(content, null, 2)
+			);
+			break;
+		}
+	}
 }
 
 export function editComment(message: any, commentController: vscode.CommentController) {
@@ -49,9 +43,9 @@ export function editComment(message: any, commentController: vscode.CommentContr
 		{ name: " " }
 	);
 	const thread = commentController.createCommentThread(
-	vscode.Uri.file(config.currentFilePath),
-	new vscode.Range(message.lineNumber - 1, 0, message.lineNumber - 1, 0),
-	[newComment]
+		vscode.Uri.file(config.currentFilePath),
+		new vscode.Range(message.lineNumber - 1, 0, message.lineNumber - 1, 0),
+		[newComment]
 	);
 	thread.canReply = false;
 	thread.collapsibleState = vscode.CommentThreadCollapsibleState.Expanded;
